Demonstrate relational reversibility issues
This is a demonstration of relations that have issues when run in
reverse (no pun intended), defining accumulator passing and multiple
direct style definitions of reverse to make the point.  Note that either
direct style definition would be perfectly reversible with concurrent
conjunction, but this is not the case with accumulator passing.